Remove IterableNodeView#flatMap - move single use case code into PathExpr.
